Decompiled a bunch of functions from 41940.c

+++ include/code0/6ACA0.h
@@ -5,7 +5,7 @@
 
 void func_8006A0A0(s16);
 void func_8006B590(s16);
-s32 func_8006D3B8(s16, s32, s32, s32, s32);
+s32 func_8006D3B8(s32, s32, s32, s32, s32);
 void func_80079560(void);
 
 /*800DF2F4*/ extern s8 D_800DF2F4[];
